What this does UPDATE
make those functions even more dynamic using our newfound knowledge of parameters and return values.

Let's update the findContact function. It should accept one parameter, index. It should now return the contact at that index.

Right now, the function is returning the contact at the second position. We'd like to call the function and find the contact at the third position, passing in the index number as an argument.

Use console.log to display the value returned by your findContact function.

Let's update the updateLastContact function. It should accept one parameter, newName. It should update the last contact in the contacts array with the newName and return the contacts array.

Then call that function and pass it the name Patty Davis.

Let's update the addContact function. It should accept one parameter, name. It should add that name to the end of the contacts array and return the contacts array.

Then call that function, passing in the name Mark Jones as a parameter.